+++ tests/ViewTest.php
@@ -19,7 +19,6 @@ use Yiisoft\View\Tests\TestSupport\TestTrait;
 use Yiisoft\View\Theme;
 use Yiisoft\View\View;
 use Yiisoft\View\ViewContextInterface;
-
 use function crc32;
 use function dechex;
 use function file_put_contents;
@@ -584,6 +583,18 @@ PHP
         $this->assertNotSame($view, $view->withClearedState());
         $this->assertNotSame($view, $view->withLocale('es'));
         $this->assertNotSame($view, $view->withFallbackExtension('tpl'));
+        $this->assertNotSame($view, $view->withTheme(new Theme([])));
+    }
+
+    public function testImmutableTheme(): void
+    {
+        $view = TestHelper::createView();
+        $theme = new Theme([]);
+        $viewWithTheme = $view->withTheme($theme);
+
+        $this->assertNull($view->getTheme());
+        $this->assertNotNull($viewWithTheme->getTheme());
+        $this->assertSame($theme, $viewWithTheme->getTheme());
     }
 
     public function testGetLocale()
@@ -604,7 +615,7 @@ PHP
 
     private function createContext(string $viewPath): ViewContextInterface
     {
-        return new class ($viewPath) implements ViewContextInterface {
+        return new class($viewPath) implements ViewContextInterface {
             public function __construct(private string $viewPath)
             {
             }