+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 8.0.1 under development
 
+- Chg: Add immutable `ViewTrait::withTheme()` method and new `ThemeState` class
 - Bug #224: Fix signature of `CachedContent::cache()` (@vjik)
 - Bug #226: Fix `reset` config for referenced definitions (@rustamwin)
 - Enh #226: Adjust config to make `View` and `WebView` more configurable (@rustamwin)